+++ src/commands/index.ts
@@ -134,7 +134,7 @@ import { docsearch } from './general/docsearch';
 import { dlss } from './support/dlss';
 import { temporarycommandedit } from './moderation/temporaryCommandEdit';
 import { temporarycommand } from './general/temporaryCommand';
-import { yourControls } from './general/yourcontrols';
+import { yourControls } from './general/yourControls';
 
 const commands: CommandDefinition[] = [
     ping,
